CheckExternalNetworks specialization for windows no longer needed
