+++ controller/internal/enforcer/applicationproxy/tcp/tcp.go
@@ -308,7 +308,7 @@ func (p *Proxy) StartClientAuthStateMachine(downIP net.IP, downPort int, downCon
 	defer downConn.SetDeadline(time.Time{}) // nolint errcheck
 
 	// First validate that L3 policies do not require a reject.
-	isEncrypted, _, err = p.CheckExternalNetwork(puContext, downIP, downPort, flowproperties, true)
+	isEncrypted, err = p.CheckExternalNetwork(puContext, downIP, downPort, flowproperties, true)
 	if err != nil {
 		return false, err
 	}
@@ -403,12 +403,9 @@ func (p *Proxy) StartServerAuthStateMachine(ip net.IP, backendport int, upConn n
 
 	// First validate that L3 policies do not require a reject.
 	//networkReport, networkPolicy, noNetAccessPolicy := puContext.NetworkACLPolicyFromAddr(upConn.RemoteAddr().(*net.TCPAddr).IP, uint16(backendport))
-	isEncrypted, externalnetwork, err := p.CheckExternalNetwork(puContext, upConn.RemoteAddr().(*net.TCPAddr).IP, backendport, flowProperties, false)
+	isEncrypted, err = p.CheckExternalNetwork(puContext, upConn.RemoteAddr().(*net.TCPAddr).IP, backendport, flowProperties, false)
 	if err != nil {
 		return false, err
-	}
-	if externalnetwork {
-		return false, nil
 	}
 
 	defer upConn.SetDeadline(time.Time{}) // nolint errcheck
@@ -481,6 +478,41 @@ func (p *Proxy) StartServerAuthStateMachine(ip net.IP, backendport int, upConn n
 			return isEncrypted, nil
 		}
 	}
+}
+
+// CompleteEndPointAuthorization -- Aporeto Handshake on top of a completed connection
+// We will define states here equivalent to SYN_SENT AND SYN_RECEIVED
+func (p *Proxy) CompleteEndPointAuthorization(downIP net.IP, downPort int, upConn, downConn net.Conn) (bool, error) {
+
+	// If the backend is not a local IP it means that we are a client.
+	if p.isLocal(upConn) {
+		return p.StartClientAuthStateMachine(downIP, downPort, downConn)
+	}
+
+	isEncrypted, err := p.StartServerAuthStateMachine(downIP, downPort, upConn)
+	if err != nil {
+		return false, err
+	}
+
+	return isEncrypted, nil
+}
+
+// CheckExternalNetwork checks if external network access is allowed
+func (p *Proxy) CheckExternalNetwork(puContext *pucontext.PUContext, IP net.IP, Port int, flowproperties *proxyFlowProperties, network bool) (bool, error) {
+	var networkReport *policy.FlowPolicy
+	var networkPolicy *policy.FlowPolicy
+	var noNetAccessPolicy error
+	if network {
+		networkReport, networkPolicy, noNetAccessPolicy = puContext.ApplicationACLPolicyFromAddr(IP, uint16(Port))
+	} else {
+		networkReport, networkPolicy, noNetAccessPolicy = puContext.NetworkACLPolicyFromAddr(IP, uint16(Port))
+
+	}
+	if noNetAccessPolicy == nil && networkPolicy.Action.Rejected() {
+		p.reportRejectedFlow(flowproperties, puContext.ManagementID(), networkPolicy.ServiceID, puContext, collector.PolicyDrop, networkReport, networkPolicy)
+		return false, fmt.Errorf("Unauthorized by Application ACLs")
+	}
+	return false, nil
 }
 
 func (p *Proxy) reportFlow(flowproperties *proxyFlowProperties, sourceID string, destID string, context *pucontext.PUContext, mode string, report *policy.FlowPolicy, actual *policy.FlowPolicy) {
